Expanding on all aliases of git
In addition to expanding only on the command git, expand on any alias
of git. For example

New-Alias g git
g co<Tab>

New-Alias got git
got co<Tab>

It does require a change to the profile. Maybe it would be better if
the decision to use the git tab expansion or the default one happened
in the posh-git stuff so there would be less need to mess with the
profile all together.