Update path to TortoiseProc to reflect rename.
TortoiseGit 1.8.0 renamed TortoiseProc to TortoiseGitProc, which caused a
bug in posh-git where calling tgit would throw an exception (could not
find TortoiseProc.exe).

This commit fixes that bug, by testing whether TortoiseGitProc exists and
using that if it does.  If it does not exist, then the default
TortoiseProc is used instead.