Stop error message when id_rsa is missing
Since we set the $Env:HOME when we launch our Git Shell, we hit 
this new code path which joins that path with the .ssh\id_rsa file.
Problem is, that file doesn't exist so we get ssh-add taking a long 
time and reporting an error.

This code path should do the same thing as the old one and return 
null if the file doesn't exist.