Basic support for Pageant as SSH agent
On PowerShell launch, check to see if GIT_SSH exists and is set to
plink. If so, check to see if pageant is running and return if it is.
If pageant is not running, attempt to start it and load all .ppk
files stored in ~/.ssh

Requires pageant to be on path for it to be launched.