Stop Get-Process from adding exceptions to $Error
Updating the Get-Process call in GitUtils.ps1 so exceptions aren't added to the global $Error array

This occurs when pageant or ssh-agent aren't running, and therefore there is no process / pid to retrieve. However this is a valid scenario so shouldn't be added to the global $Error array.

Also being more explicit with the check of the returned value from Get-Process