Add a flag to suppress deprecation warning
Allow Powershell 2 support deprecation warning to be suppressed by
importing the module with a true value as an argument.

The ability to pass arguments through Import-Module is pretty limited,
accepting only positional arguments, but it still seems a preferable
option to environment variables or asking users to modify posh-git
directly.

Closes #163