Export Write-VcsStatus to help module autoloading
By exporting `Write-VcsStatus` we can just check if we're in a git directory, only then call this function to style the prompt. Having powershell load the module into the session. Doing it at the start slows the startup for no reason if the user isn't in a git directory.

Seems you only have to mention the export in the .psd1 file for it to autoload.