Adds help topic for module.
If you ship this then users of posh-git will be able to execute "man about_posh-git" and get the info that is currently in the readme.md file.  It is recommended that PowerShell modules provide an about help topic for users to get started with using a module.  It is also a very handy way looking up the "legend" for the status prompt.