WIP allow prompt char > to reflect nesting level.
To be a good prompt citizen in PowerShell, the > (or whatever prompt char) should be displayed once for nesting level 0, twice for nesting level 1, and so on.

In order to make that happen, we have to allow for deferred expansion of the Prompt*Suffix settings.

See https://msdn.microsoft.com/en-us/powershell/reference/5.1/microsoft.powershell.core/about/about_prompts for more details.