Worth adding optional timing output to def prompt?
This is enabled with `$GitPromptSettings.DefaultPromptEnableTiming=$true` and then the prompt function outputs `~\GitHub\rkeithhill\posh-git [rkeithhill/enable-default-prompt-timing +0 ~2 -0 !] 55 mS>`.

It also occurred to me that the GitPromptSettings should be prefixed "DefaultPrompt" so folks don't think they apply to their own custom prompt functions.

I'm not entirely sure this is worth it but I could see it coming in handy to get "hard" perf data from users of the default prompt.  If  we decide not to go with this PR, I still think we should consider adding the "Default" prefix to the existing two Prompt*Suffix settings.

In the VSCode dir on my machine (status equiv to origin), the prompt takes 150 mS.  VSCode has 6466 files in the repo.