Redo .git dir look up.
If we find a file named HEAD and a dir named refs then we use rev-parse.  Updated Invoke-Utf8ConsoleCommand to clear out any errors that it caused.

Not done.  Still need to take those errors and put them in a module variable but this will be useful for testing if we want to go this route.