SignatureType enum is not available on PS v2
We now rely on the SignerCertificate field to be $null.  If it isn't $null, then we assume the profile is signed and we don't modify it.  Note that this field is not $null when the file is signed but there is a hash mismatch (which is the desired behavior). Fix #448.