Add vsts.pr as a "known alias"
The idea "known alias" is that we can standardize handling of a
particular alias command (e.g. 'git pr' pass-through to 'vsts code pr')
without hooking on the alias name. For example, 'pr' could reasonably be
used to manage GitHub, so this allows us to avoid using vsts tab
expansion in that case.

Namespacing the alias ('vsts.pr') prevents conflicts with a real alias.