Modifying tests to return paths with leading "./".
Resolve-Path always renders files with a leading ".\" (Windows) or "./" (everything else).
This differs from how paths were returned before, so now the tests are changing to reflect this.
It is also now consistent with how tab completion works for other PowerShell commands (e.g., Get-ChildItem), so it shouldn't be a problem.