Make SSH_AGENT_PID mapping to Windows PID
Port of https://github.com/dahlbyk/posh-git/pull/699. This does drop -ErrorAction Ignore as unnecessary. It further adds [0] to full name parent directory for bin / usr lookup of ssh location as the lookup of 'ps' and and will get multiple (2) results. This ensures it only uses the first in that situation.